requirements.txt
taipy-gui>=3.0.0
pandas>=2.0.0
networkx>=3.0
plotly>=5.0.0
matplotlib